Test2/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
Pillow>=10.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
pymilvus>=2.3.0
pydantic>=2.0.0
efficientnet-pytorch>=0.7.0
